added userId check on delete, getAll and getSingle

+++ Services/CharacterService/CharacterService.cs
@@ -1,4 +1,5 @@
-﻿using AutoMapper;
+﻿using System.Security.Claims;
+using AutoMapper;
 using Microsoft.EntityFrameworkCore;
 using rpg_backend.Data;
 using rpg_backend.Dtos.Character;
@@ -9,22 +10,28 @@ public class CharacterService : ICharacterService
 {
     private readonly IMapper _mapper;
     private readonly DataContext _context;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
         _mapper = mapper;
         _context = context;
+        _httpContextAccessor = httpContextAccessor;
     }
 
-    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters(int userId)
+    private int GetUserId() =>
+        int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
+
+    public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
         var response = new ServiceResponse<List<GetCharacterDto>>();
-        var dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+        var dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
         if (dbCharacters.Count == 0)
         {
             response.Message = "No characters found";
-            return response;    
+            return response;
         }
+
         response.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
         return response;
     }
@@ -32,7 +39,7 @@ public class CharacterService : ICharacterService
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
         var serviceResponse = new ServiceResponse<GetCharacterDto>();
-        var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+        var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
         serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
         return serviceResponse;
     }
@@ -41,12 +48,14 @@ public class CharacterService : ICharacterService
     {
         var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
         Character character = _mapper.Map<Character>(newCharacter);
+        character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
         //no need for async as we are not waiting for the result
         _context.Characters.Add(character);
         await _context.SaveChangesAsync();
-        serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+        serviceResponse.Data = await _context.Characters.Where(c => c.User.Id == GetUserId())
+            .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
         serviceResponse.Message = "Character added successfully";
-        
+
         return serviceResponse;
     }
 
@@ -54,7 +63,7 @@ public class CharacterService : ICharacterService
     {
         ServiceResponse<GetCharacterDto> response = new ServiceResponse<GetCharacterDto>();
 
-        try 
+        try
         {
             var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
             // character.Name = updatedCharacter.Name;
@@ -84,11 +93,20 @@ public class CharacterService : ICharacterService
 
         try
         {
-            Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-            _context.Characters.Remove(character);
-            await _context.SaveChangesAsync();
-
-            response.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            Character character =
+                await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+            if (character != null)
+            {
+                _context.Characters.Remove(character);
+                await _context.SaveChangesAsync();
+                response.Data = _context.Characters.Where(c => c.User.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+            }
+            else
+            {
+                response.Success = false;
+                response.Message = "Character not found";
+            }
         }
         catch (Exception ex)
         {
